pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Ranjan Kumar <ranjandeo.gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openpyxl = "^3.1.5"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:main"
